[REF] shopfloor_reception: adapt test suite to "set_lot" refactoring
